react router dom upgraded and migrating from BrowserRouter component to RouterProvider
