+++ package.json
@@ -17,7 +17,7 @@
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-fela": "^11.7.0",
-    "react-router-dom": "^6.2.1",
+    "react-router-dom": "^6.24.0",
     "react-scripts": "5.0.0",
     "recharts": "^2.1.9",
     "sass": "^1.49.8",
