+++ src/index.js
@@ -1,13 +1,13 @@
 // import React from 'react';
 import "./sass/main.scss";
 import ReactDOM from "react-dom";
-import App from "./App";
 import { GlobalContext } from "./context/GlobalContext";
 import { RendererProvider, ThemeProvider } from "react-fela";
 import theme from "./felaConfig/theme";
-import { BrowserRouter } from "react-router-dom";
+// import { BrowserRouter } from "react-router-dom";
 import { createRenderer } from "fela";
 import { ApolloClient, InMemoryCache, ApolloProvider } from "@apollo/client";
+import App from "./App";
 
 const renderer = createRenderer();
 // console.log(process.env.REACT_APP_API_BASE);
@@ -22,15 +22,15 @@ const client = new ApolloClient({
 
 ReactDOM.render(
   <ApolloProvider client={client}>
-    <BrowserRouter>
-      <RendererProvider renderer={renderer}>
-        <ThemeProvider theme={theme}>
-          <GlobalContext>
-            <App />
-          </GlobalContext>
-        </ThemeProvider>
-      </RendererProvider>
-    </BrowserRouter>
+    {/* <BrowserRouter> */}
+    <RendererProvider renderer={renderer}>
+      <ThemeProvider theme={theme}>
+        <GlobalContext>
+          <App />
+        </GlobalContext>
+      </ThemeProvider>
+    </RendererProvider>
+    {/* </BrowserRouter> */}
   </ApolloProvider>,
   document.getElementById("root")
 );
